finish all schemas except forum

+++ src/main/java/com/kttdevelopment/myanimelist/property/ExtendedPreview.java
@@ -14,7 +14,7 @@ public interface ExtendedPreview extends Preview {
 
     String getSynopsis();
 
-    int getMeanRating();
+    double getMeanRating();
 
     int getRank();
